Encapsulated alch functions within alch menu
