+++ client/src/App.jsx
@@ -14,17 +14,11 @@ import creatures from "./constants/creatures";
 import relics from "./constants/relics";
 import { potionsList } from "./constants/items";
 import { ingredientsList } from "./constants/items";
-import recipeList from "./constants/recipes";
-import { getItems, addItem } from "./services/itemServices";
-import { getPotionTimer, addPotionTimer } from "./services/potionTimerServices";
+import { getItems } from "./services/itemServices";
 import { enemyCreaturesHome } from "./constants/enemyCreatures";
 import { useSelector, useDispatch } from "react-redux";
 import { setPlayerCreatureValue } from "./store/actions/summon.actions";
 import {
-  enablePotionCooldown,
-  disablePotionCooldown,
-  setSummonHPBonusAmount,
-  setSummonMPBonusAmount,
   setIngredientsValue,
   setPotionsValue,
 } from "./store/actions/alchemy.actions";
@@ -44,12 +38,6 @@ function App() {
 
   // player creature state from redux store
   const playerCreature = useSelector((state) => state.summon.playerCreature);
-  // relics state from redux store
-  const chosenRelic = useSelector((state) => state.relics.chosenRelic);
-  // alchemy state from redux store
-  const summonHPBonus = useSelector((state) => state.alchemy.summonHPBonus);
-  const summonMPBonus = useSelector((state) => state.alchemy.summonMPBonus);
-  const potionCooldown = useSelector((state) => state.alchemy.potionCooldown);
 
   // navigation hook
   const navigate = useNavigate();
@@ -260,185 +248,6 @@ function App() {
     }
   };
 
-  // creates a new potion
-  const createPotion = async (potionId) => {
-    try {
-      if (!potionCooldown) {
-        dispatch(enablePotionCooldown());
-        const { data } = await getItems();
-        const playerPotionData = data.filter(
-          (item) => item.type === "Potion" && item.itemId === potionId
-        );
-        const potion = potionsList.find((item) => item.id === potionId);
-        const newPotionData = {
-          itemId: potion.id,
-          type: "Potion",
-          itemQuantity: playerPotionData[0]
-            ? playerPotionData[0].itemQuantity + 1
-            : 1,
-          userId: Userfront.user.userId,
-        };
-        const playerIngredientData = data.filter(
-          (item) => item.type === "Ingredient"
-        );
-        const currentRecipe = recipeList.filter(
-          (item) => item.potionProductId === potion.id
-        )[0];
-        // check if player has enough ingredients for recipe
-        const ingredient1Check = playerIngredientData.find(
-          (item) => item.itemId === currentRecipe.ingredient1
-        );
-        const ingredient2Check = playerIngredientData.find(
-          (item) => item.itemId === currentRecipe.ingredient2
-        );
-        if (
-          ingredient1Check &&
-          ingredient1Check.itemQuantity > 0 &&
-          ingredient2Check &&
-          ingredient2Check.itemQuantity > 0
-        ) {
-          // confirm potion creation
-          if (window.confirm(`Are you sure you want to create this potion?`)) {
-            const currentIngredient1 = playerIngredientData.find(
-              (ingredient) => ingredient.itemId === currentRecipe.ingredient1
-            );
-            const currentIngredient2 = playerIngredientData.find(
-              (ingredient) => ingredient.itemId === currentRecipe.ingredient2
-            );
-            // delete the ingredients from the player's inventory
-            currentIngredient1.itemQuantity -= 1;
-            currentIngredient2.itemQuantity -= 1;
-            await Userfront.user.update({
-              data: {
-                userkey: Userfront.user.data.userkey,
-              },
-            });
-            await addItem(currentIngredient1);
-            await loadDataAlchemy();
-            await Userfront.user.update({
-              data: {
-                userkey: Userfront.user.data.userkey,
-              },
-            });
-            await addItem(currentIngredient2);
-            await loadDataAlchemy();
-            // add the potion to the player's inventory
-            await Userfront.user.update({
-              data: {
-                userkey: Userfront.user.data.userkey,
-              },
-            });
-            await addItem(newPotionData);
-            await loadDataAlchemy();
-            setTimeout(() => {
-              dispatch(disablePotionCooldown());
-            }, 1000);
-          } else {
-            setTimeout(() => {
-              dispatch(disablePotionCooldown());
-            }, 1000);
-          }
-        } else {
-          alert("You don't have enough ingredients for this potion.");
-          setTimeout(() => {
-            dispatch(disablePotionCooldown());
-          }, 1000);
-        }
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  // uses clicked potion
-  const consumePotion = async (potionId) => {
-    try {
-      if (!potionCooldown) {
-        await getPotionTimer();
-        const timerData = await getPotionTimer();
-        dispatch(enablePotionCooldown());
-        const { data } = await getItems();
-        const playerPotionData = data.filter(
-          (item) => item.type === "Potion" && item.itemId === potionId
-        );
-        const potion = potionsList.find((item) => item.id === potionId);
-        const currentPotionData = playerPotionData.filter(
-          (item) => item.itemId === potionId
-        );
-        // prevent over use
-        if (timerData.data.length !== 0) {
-          alert(
-            "You already have an active potion. Please wait for it to expire."
-          );
-          setTimeout(() => {
-            dispatch(disablePotionCooldown());
-          }, 1000);
-          return;
-        }
-        // check if player has enough potions
-        if (currentPotionData[0].itemQuantity > 0) {
-          // confirm potion use
-          if (
-            window.confirm(`Are you sure you want to use this potion?`) === true
-          ) {
-            // use potion
-            currentPotionData[0].itemQuantity -= 1;
-            await Userfront.user.update({
-              data: {
-                userkey: Userfront.user.data.userkey,
-              },
-            });
-            await addItem(currentPotionData[0]);
-            await loadDataAlchemy();
-            await Userfront.user.update({
-              data: {
-                userkey: Userfront.user.data.userkey,
-              },
-            });
-            await addPotionTimer({
-              userId: Userfront.user.userId,
-              potionId: potion.id,
-              potionDuration: potion.duration,
-            });
-            await loadDataAlchemy();
-            // checks potion timer
-            const potionTimer = await getPotionTimer();
-            if (potionTimer.data.length > 0) {
-              const playerPotion = potionsList.find(
-                (potion) => potion.id === potionTimer.data[0].potionId
-              );
-              const playerMPBonus = playerPotion.mpMod;
-              const playerHPBonus = playerPotion.hpMod;
-              dispatch(setSummonMPBonusAmount(playerMPBonus));
-              dispatch(setSummonHPBonusAmount(playerHPBonus));
-            }
-
-            setPlayerCreatureMP(
-              playerCreature.mp + chosenRelic.mpMod + summonMPBonus
-            );
-            setPlayerCreatureHP(
-              playerCreature.hp + chosenRelic.hpMod + summonHPBonus
-            );
-            setTimeout(() => {
-              dispatch(disablePotionCooldown());
-            }, 1000);
-          } else {
-            setTimeout(() => {
-              dispatch(disablePotionCooldown());
-            }, 1000);
-          }
-        } else {
-          alert("You don't have enough potions.");
-          setTimeout(() => {
-            dispatch(disablePotionCooldown());
-          }, 1000);
-        }
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
   // renders if a player creature and relic is bestowed
   if (playerCreature && player.chosenRelic) {
     return (
@@ -491,10 +300,13 @@ function App() {
               />
 
               <AlchemyMenu
+                Userfront={Userfront}
                 gameMenuStatus={gameMenuStatus}
                 setGameMenuStatus={setGameMenuStatus}
-                createPotion={createPotion}
-                consumePotion={consumePotion}
+                loadDataAlchemy={() => loadDataAlchemy()}
+                playerCreature={playerCreature}
+                setPlayerCreatureHP={setPlayerCreatureHP}
+                setPlayerCreatureMP={setPlayerCreatureMP}
               />
 
               <PlayerCreature
